// convex/_generated/api.d.ts
/* eslint-disable */
/**
 * Generated `api` utility.
 *
 * THIS CODE IS AUTOMATICALLY GENERATED.
 *
 * To regenerate, run `npx convex dev`.
 * @module
 */

import type {
  ApiFromModules,
  FilterApi,
  FunctionReference,
} from "convex/server";
import type * as auth from "../auth.js";
import type * as evaluations from "../evaluations.js";
import type * as feedback from "../feedback.js";
import type * as panels from "../panels.js";
import type * as questions from "../questions.js";
import type * as sessions from "../sessions.js";
import type * as skills from "../skills.js";
import type * as users from "../users.js";

/**
 * A utility for referencing Convex functions in your app's API.
 *
 * Usage:
 * ```js
 * const myFunctionReference = api.myModule.myFunction;
 * ```
 */
declare const fullApi: ApiFromModules<{
  auth: typeof auth;
  evaluations: typeof evaluations;
  feedback: typeof feedback;
  panels: typeof panels;
  questions: typeof questions;
  sessions: typeof sessions;
  skills: typeof skills;
  users: typeof users;
}>;
export declare const api: FilterApi<
  typeof fullApi,
  FunctionReference<any, "public">
>;
export declare const internal: FilterApi<
  typeof fullApi,
  FunctionReference<any, "internal">
>;